Add support of Force boolean flag
Force boolean (false|true) flag has been added in OData user synchronizer
since DHuS V0.9.0-hotfix-3. The Force flag aims to perform synchronization
of users with username only when true. Otherwise, the matching of the user
if performed according to the username and the date of creation of the user.